Publish the post on Rails async db queries
The post is also accompained by a gist:
https://gist.github.com/radanskoric/48f1982f2fe80b7d3bb44680f6d292aa

In addition, while working on the post I realised that
I lost the convertkit signup when switching to Chirpy theme.
I restored it and for this article I also inserted a signup
half way through the article.

Because convertkit out of the box only looks good on the light
theme I decided to turn off light/dark switcher for now. I tried
getting to adapt but decided it's not a good use of my time now.
I logged a task to get back to it in the future.